graphql: make sure gqlV1 and gql are not used in the same files
