+++ components/edit-collective/sections/Policies.js
@@ -8,7 +8,8 @@ import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
 import { MODERATION_CATEGORIES } from '../../../lib/constants/moderation-categories';
 import { i18nGraphqlException } from '../../../lib/errors';
 import { DEFAULT_SUPPORTED_EXPENSE_TYPES } from '../../../lib/expenses';
-import { API_V2_CONTEXT, gql, gqlV1 } from '../../../lib/graphql/helpers';
+import { API_V2_CONTEXT, gql } from '../../../lib/graphql/helpers';
+import { editCollectivePolicyMutation } from '../../../lib/graphql/v1/mutations';
 import useLoggedInUser from '../../../lib/hooks/useLoggedInUser';
 import { stripHTML } from '../../../lib/html';
 import { PREVIEW_FEATURE_KEYS } from '../../../lib/preview-features';
@@ -38,17 +39,6 @@ const CONTRIBUTION_POLICY_MAX_LENGTH = 3000; // 600 words * 5 characters average
 const updateFilterCategoriesMutation = gql`
   mutation UpdateFilterCategories($account: AccountReferenceInput!, $key: AccountSettingsKey!, $value: JSON!) {
     editAccountSetting(account: $account, key: $key, value: $value) {
-      id
-      type
-      isActive
-      settings
-    }
-  }
-`;
-
-const editCollectiveMutation = gqlV1/* GraphQL */ `
-  mutation EditCollective($collective: CollectiveInputType!) {
-    editCollective(collective: $collective) {
       id
       type
       isActive
@@ -156,7 +146,7 @@ const Policies = ({ collective, showOnlyExpensePolicy }) => {
     },
   );
   const [updateCollective, { loading: isSubmittingSettings, error: settingsError }] =
-    useMutation(editCollectiveMutation);
+    useMutation(editCollectivePolicyMutation);
   const [setPolicies, { loading: isSettingPolicies, error: policiesError }] = useMutation(setPoliciesMutation, {
     context: API_V2_CONTEXT,
   });
