+++ components/edit-collective/sections/SendingMoney.js
@@ -4,7 +4,7 @@ import { graphql } from '@apollo/client/react/hoc';
 import { FormattedMessage, injectIntl } from 'react-intl';
 
 import hasFeature, { FEATURES } from '../../../lib/allowed-features';
-import { gqlV1 } from '../../../lib/graphql/helpers';
+import { editCollectiveSettingsMutation } from '../../../lib/graphql/v1/mutations';
 
 import MessageBox from '../../MessageBox';
 import StyledButton from '../../StyledButton';
@@ -110,15 +110,6 @@ class SendingMoney extends React.Component {
     );
   }
 }
-
-const editCollectiveSettingsMutation = gqlV1/* GraphQL */ `
-  mutation EditCollectiveSettings($id: Int!, $settings: JSON) {
-    editCollective(collective: { id: $id, settings: $settings }) {
-      id
-      settings
-    }
-  }
-`;
 
 const addEditCollectiveSettingsMutation = graphql(editCollectiveSettingsMutation, {
   name: 'editCollectiveSettings',
