+++ components/onboarding-modal/OnboardingModal.js
@@ -9,8 +9,9 @@ import styled, { css } from 'styled-components';
 
 import { confettiFireworks } from '../../lib/confettis';
 import { getErrorFromGraphqlException } from '../../lib/errors';
-import { API_V2_CONTEXT, gql, gqlV1 } from '../../lib/graphql/helpers';
+import { API_V2_CONTEXT, gql } from '../../lib/graphql/helpers';
 import { SocialLinkType } from '../../lib/graphql/types/v2/graphql';
+import { editCollectiveContactMutation, editCollectiveMembersMutation } from '../../lib/graphql/v1/mutations';
 import { compose, isValidUrl } from '../../lib/utils';
 
 import Container from '../../components/Container';
@@ -390,39 +391,9 @@ class OnboardingModal extends React.Component {
   }
 }
 
-// GraphQL for editing Collective admins info
-const editCollectiveMembersMutation = gqlV1/* GraphQL */ `
-  mutation EditCollectiveMembers($collectiveId: Int!, $members: [MemberInputType!]!) {
-    editCoreContributors(collectiveId: $collectiveId, members: $members) {
-      id
-      members(roles: ["ADMIN"]) {
-        id
-        role
-        member {
-          id
-          name
-        }
-      }
-    }
-  }
-`;
-
 export const addEditCollectiveMembersMutation = graphql(editCollectiveMembersMutation, {
   name: 'editCollectiveMembers',
 });
-
-// GraphQL for editing Collective contact info
-const editCollectiveContactMutation = gqlV1/* GraphQL */ `
-  mutation EditCollectiveContact($collective: CollectiveInputType!) {
-    editCollective(collective: $collective) {
-      id
-      socialLinks {
-        type
-        url
-      }
-    }
-  }
-`;
 
 const addEditCollectiveContactMutation = graphql(editCollectiveContactMutation, {
   name: 'editCollectiveContact',
