+++ components/root-actions/UnhostAccountForm.js
@@ -3,33 +3,13 @@ import { useMutation } from '@apollo/client';
 import { useIntl } from 'react-intl';
 
 import { i18nGraphqlException } from '../../lib/errors';
-import { API_V2_CONTEXT, gql, gqlV1 } from '../../lib/graphql/helpers';
+import { API_V2_CONTEXT, gql } from '../../lib/graphql/helpers';
+import { unhostAccountCollectivePickerSearchQuery } from '../../lib/graphql/v1/queries';
 
 import CollectivePickerAsync from '../CollectivePickerAsync';
 import StyledButton from '../StyledButton';
 import StyledInputField from '../StyledInputField';
 import { useToast } from '../ui/useToast';
-
-const collectivePickerSearchQuery = gqlV1/* GraphQL */ `
-  query UnhostAccountSearch($term: String!, $types: [TypeOfCollective], $limit: Int, $hostCollectiveIds: [Int]) {
-    search(term: $term, types: $types, limit: $limit, hostCollectiveIds: $hostCollectiveIds) {
-      id
-      collectives {
-        id
-        type
-        slug
-        name
-        imageUrl(height: 64)
-        isActive
-        host {
-          id
-          slug
-          name
-        }
-      }
-    }
-  }
-`;
 
 const unhostAccountMutation = gql`
   mutation UnhostAccount($account: AccountReferenceInput!) {
@@ -59,7 +39,7 @@ const UnhostAccountForm = () => {
             inputId={id}
             onChange={({ value }) => setAccount(value)}
             types={['COLLECTIVE', 'FUND']}
-            searchQuery={collectivePickerSearchQuery}
+            searchQuery={unhostAccountCollectivePickerSearchQuery}
             filterResults={collectives => collectives.filter(c => Boolean(c.host))}
             collective={account}
             noCache
