+++ lib/graphql/v1/fragments.js
@@ -1,0 +1,8 @@
+import { gqlV1 } from '../helpers';
+
+export const collectiveBalanceFragment = gqlV1/* GraphQL */ `
+  fragment StatFieldsFragment on CollectiveStatsType {
+    id
+    balance
+  }
+`;
