+++ lib/hooks/withLoggedInUser.js
@@ -5,7 +5,7 @@ import dayjs from 'dayjs';
 import { decodeJwt } from 'jose';
 
 import { exchangeLoginToken, refreshToken, refreshTokenWithTwoFactorCode } from '../api';
-import { loggedInUserQuery } from '../graphql/queries';
+import { loggedInUserQuery } from '../graphql/v1/queries';
 import { getFromLocalStorage, LOCAL_STORAGE_KEYS, setLocalStorage } from '../local-storage';
 import LoggedInUser from '../LoggedInUser';
