+++ pages/banner-iframe.js
@@ -8,7 +8,8 @@ import { FormattedMessage } from 'react-intl';
 import styled from 'styled-components';
 
 import { CollectiveType } from '../lib/constants/collectives';
-import { API_V2_CONTEXT, gql, gqlV1 } from '../lib/graphql/helpers';
+import { API_V2_CONTEXT, gql } from '../lib/graphql/helpers';
+import { collectiveBannerIframeQuery } from '../lib/graphql/v1/queries';
 import { getRequestIntl } from '../lib/i18n/request';
 import { parseToBoolean } from '../lib/utils';
 
@@ -429,26 +430,6 @@ class BannerIframe extends React.Component {
     );
   }
 }
-
-const collectiveBannerIframeQuery = gqlV1/* GraphQL */ `
-  query CollectiveBannerIframe($collectiveSlug: String) {
-    Collective(slug: $collectiveSlug) {
-      id
-      name
-      slug
-      currency
-      stats {
-        id
-        backers {
-          id
-          users
-          organizations
-          collectives
-        }
-      }
-    }
-  }
-`;
 
 export const addCollectiveBannerIframeData = graphql(collectiveBannerIframeQuery, {
   options({ collectiveSlug, useNewFormat }) {
